feat(data-classes): some requestContext refactoring
