+++ tests/functional/test_data_classes.py
@@ -792,6 +792,7 @@ def test_default_api_gateway_proxy_event():
     assert identity.user == event["requestContext"]["identity"]["user"]
     assert identity.user_agent == event["requestContext"]["identity"]["userAgent"]
     assert identity.user_arn == event["requestContext"]["identity"]["userArn"]
+    assert identity.client_cert.subject_dn == "www.example.com"
 
     assert request_context.path == event["requestContext"]["path"]
     assert request_context.protocol == event["requestContext"]["protocol"]
@@ -858,6 +859,7 @@ def test_api_gateway_proxy_event():
     assert identity.user == event["requestContext"]["identity"]["user"]
     assert identity.user_agent == event["requestContext"]["identity"]["userAgent"]
     assert identity.user_arn == event["requestContext"]["identity"]["userArn"]
+    assert identity.client_cert.subject_dn == "www.example.com"
 
     assert request_context.path == event["requestContext"]["path"]
     assert request_context.protocol == event["requestContext"]["protocol"]
@@ -1546,6 +1548,42 @@ def test_api_gateway_authorizer_request_event():
     assert event.query_string_parameters == event["queryStringParameters"]
     assert event.path_parameters == event["pathParameters"]
     assert event.stage_variables == event["stageVariables"]
+
+    assert event.request_context is not None
+    request_context = event.request_context
+    assert request_context.account_id == event["requestContext"]["accountId"]
+    assert request_context.api_id == event["requestContext"]["apiId"]
+
+    assert request_context.domain_name == event["requestContext"]["domainName"]
+    assert request_context.domain_prefix == event["requestContext"]["domainPrefix"]
+    assert request_context.extended_request_id == event["requestContext"]["extendedRequestId"]
+    assert request_context.http_method == event["requestContext"]["httpMethod"]
+
+    identity = request_context.identity
+    assert identity.access_key == event["requestContext"]["identity"]["accessKey"]
+    assert identity.account_id == event["requestContext"]["identity"]["accountId"]
+    assert identity.caller == event["requestContext"]["identity"]["caller"]
+    assert (
+        identity.cognito_authentication_provider == event["requestContext"]["identity"]["cognitoAuthenticationProvider"]
+    )
+    assert identity.cognito_authentication_type == event["requestContext"]["identity"]["cognitoAuthenticationType"]
+    assert identity.cognito_identity_id == event["requestContext"]["identity"]["cognitoIdentityId"]
+    assert identity.cognito_identity_pool_id == event["requestContext"]["identity"]["cognitoIdentityPoolId"]
+    assert identity.principal_org_id == event["requestContext"]["identity"]["principalOrgId"]
+    assert identity.source_ip == event["requestContext"]["identity"]["sourceIp"]
+    assert identity.user == event["requestContext"]["identity"]["user"]
+    assert identity.user_agent == event["requestContext"]["identity"]["userAgent"]
+    assert identity.user_arn == event["requestContext"]["identity"]["userArn"]
+    assert identity.client_cert.subject_dn == "www.example.com"
+
+    assert request_context.path == event["requestContext"]["path"]
+    assert request_context.protocol == event["requestContext"]["protocol"]
+    assert request_context.request_id == event["requestContext"]["requestId"]
+    assert request_context.request_time == event["requestContext"]["requestTime"]
+    assert request_context.request_time_epoch == event["requestContext"]["requestTimeEpoch"]
+    assert request_context.resource_id == event["requestContext"]["resourceId"]
+    assert request_context.resource_path == event["requestContext"]["resourcePath"]
+    assert request_context.stage == event["requestContext"]["stage"]
 
 
 def test_api_gateway_authorizer_simple_response():
